ml/requirements.txt
langchain
langgraph
langchain_community
langchain_openai
langchain_core
pydantic
typing_extensions
typing
tiktoken
chromadb
langchainhub
pypdf
pathway
openparse
python-dotenv
torch
torchvision
transformers
langsmith
ragas
polars
openparse
cohere
gradio
numexpr
e2b_code_interpreter
anthropic
instructor
langchain-google-community
bs4
pdfkit
